+++ modules/tinymce/src/core/test/ts/browser/content/InsertContentTest.ts
@@ -351,4 +351,83 @@ describe('browser.tinymce.core.content.InsertContentTest', () => {
     InsertContent.insertAtCaret(editor, ' <strong> a </strong> ');
     TinyAssertions.assertContent(editor, '<pre> <strong> a </strong> </pre>');
   });
+
+  it('TINY-6263: insertAtCaret - merge font-size spans', () => {
+    const editor = hook.editor();
+    editor.setContent('');
+    TinySelections.setCursor(editor, [ 0 ], 0);
+    InsertContent.insertAtCaret(editor, {
+      content: '<p>' +
+        '<span style="font-size: 9pt;">' +
+        '<span style="font-size: 14pt;">' +
+        '<span style="font-size: 9pt;">' +
+        '<span style="font-size: 9pt;">test</span>' +
+        '</span>' +
+        '</span>' +
+        '</span>' +
+        '</p>',
+      merge: true
+    });
+    TinyAssertions.assertContent(editor, '<p>' +
+      '<span style="font-size: 9pt;">' +
+      '<span style="font-size: 14pt;">' +
+      '<span style="font-size: 9pt;">test</span>' +
+      '</span>' +
+      '</span>' +
+      '</p>');
+  });
+
+  it('TINY-6263: insertAtCaret - merge spans with similar node in-between', () => {
+    const editor = hook.editor();
+    editor.setContent('');
+    TinySelections.setCursor(editor, [ 0 ], 0);
+    InsertContent.insertAtCaret(editor, {
+      content: '<p>' +
+        '<span style="color: red; font-size: 9pt;">' +
+        '<span style="background-color: red; color: red;">' +
+        '<span style="color: red; font-size: 9pt;">test</span>' +
+        '</span>' +
+        '</span>' +
+        '</p>',
+      merge: true
+    });
+    TinyAssertions.assertContent(editor, '<p>' +
+      '<span style="color: red; font-size: 9pt;">' +
+      '<span style="background-color: red; color: red;">test</span>' +
+      '</span>' +
+      '</p>');
+  });
+
+  it('TINY-6263: insertAtCaret - merge font colors with other surrounding inline elements in-between', () => {
+    const editor = hook.editor();
+    editor.setContent('');
+    TinySelections.setCursor(editor, [ 0 ], 0);
+    InsertContent.insertAtCaret(editor, {
+      content: '<p>' +
+        '<span style="color: yellow;">' +
+        '<span style="background-color: red;">' +
+        '<span style="color: yellow;">' +
+        '<span style="color: red;">red</span>' +
+        'yellow' +
+        '<span style="color: blue;">' +
+        '<strong>' +
+        '<span style="color: blue;">blue</span>' +
+        '</strong>' +
+        '</span>' +
+        '</span>' +
+        '</span>' +
+        '</span>' +
+        '</p>',
+      merge: true
+    });
+    TinyAssertions.assertContent(editor, '<p>' +
+    '<span style="color: yellow;">' +
+    '<span style="background-color: red;">' +
+    '<span style="color: red;">red</span>' +
+    'yellow' +
+    '<span style="color: blue;"><strong>blue</strong></span>' +
+    '</span>' +
+    '</span>' +
+    '</p>');
+  });
 });